+++ scripts/events/abilityCheck.js
@@ -1,6 +1,7 @@
 import {DialogApp} from '../applications/dialog.js';
 import {custom} from './custom.js';
 import {actorUtils, effectUtils, genericUtils, itemUtils, macroUtils, regionUtils, templateUtils} from '../utils.js';
+import {heroicInspiration} from '../macros/2024/mechanics/heroicInspiration.js';
 function getMacroData(entity) {
     return entity.flags['chris-premades']?.macros?.check ?? [];
 }
@@ -198,10 +199,9 @@ function _applyDeprecatedD20Configs(rollConfig, dialogConfig, messageConfig, opt
     set(messageConfig, 'data', options.messageData);
     set(messageConfig, 'rollMode', options.rollMode);
     set(messageConfig, 'data.flavor', options.flavor);
-  
-    if ( !foundry.utils.isEmpty(roll) ) {
+    if (!foundry.utils.isEmpty(roll)) {
         rollConfig.rolls ??= [];
-        if ( rollConfig.rolls[0] ) rollConfig.rolls[0] = roll;
+        if ( rollConfig.rolls[0]) rollConfig.rolls[0] = roll;
         else rollConfig.rolls.push(roll);
     }
 }
@@ -313,6 +313,10 @@ async function rollCheck(wrapped, config, dialog = {}, message = {}) {
             let bonusRoll = await executeMacro(trigger);
             if (bonusRoll) returnData = CONFIG.Dice.D20Roll.fromRoll(bonusRoll);
         }
+    }
+    if (genericUtils.getCPRSetting('heroicInspiration')) {
+        let heroicInspirationRoll = await heroicInspiration.saveSkillCheck(returnData, this);
+        if (heroicInspirationRoll) returnData = heroicInspirationRoll;
     }
     if (returnData.options) genericUtils.mergeObject(returnData.options, oldOptions);
     if (message.create !== false) {