+++ scripts/settings.js
@@ -726,6 +726,12 @@ export function registerSettings() {
         category: 'mechanics'
     });
     addSetting({
+        key: 'heroicInspiration',
+        type: Boolean,
+        default: false,
+        category: 'mechanics'
+    });
+    addSetting({
         key: 'bg3WeaponActionsEnabled',
         type: Boolean,
         default: false,